+++ src/commands/helps/mod.rs
@@ -5,6 +5,7 @@ pub mod env;
 pub mod get;
 pub mod init;
 pub mod install;
+pub mod launch;
 pub mod lock;
 pub mod new;
 pub mod orbit;