+++ src/commands/manuals/mod.rs
@@ -9,7 +9,6 @@ pub mod launch;
 pub mod lock;
 pub mod new;
 pub mod orbit;
-pub mod plan;
 pub mod read;
 pub mod remove;
 pub mod search;